base/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
attrs==24.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
ecdsa==0.19.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
